Modularize vaues
